♻️ Refactor: 영수증의 format을 맞출 때, 사용하는 text 길이 함수를 textUtils.ts로 분리
